change: rename api directory
